refactor(component): 상품 관련 공통 컴포넌트 폴더 분리
