+++ app/(routes)/(main)/_components/SpecialProductsSection.tsx
@@ -1,4 +1,4 @@
-import HighlightedProductsCarousel from '@/app/_components/shared/HighlightedProductsCarousel';
+import HighlightedProductsCarousel from '@/app/_components/product/HighlightedProductsCarousel';
 import { productListMock } from '@/lib/mocks/product-list.mock';
 
 interface SpecialProductsSectionProps {
