+++ app/(routes)/products/page.tsx
@@ -1,0 +1,3 @@
+export default function Products() {
+  return <div>상품 목록 페이지 (종합)</div>;
+}
